meandthebois/BudgetBuddy_api_test/requirements.txt
aiohttp==3.9.3
aiosignal==1.3.1
asgiref==3.7.2
attrs==23.1.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
cryptography==42.0.2
defusedxml==0.8.0rc2
dj-database-url==2.1.0
Django==4.2.9
django-cors-headers==4.3.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
frozenlist==1.4.1
gunicorn==21.2.0
h11==0.14.0
idna==3.4
multidict==6.0.4
oauthlib==3.2.2
openai==0.28.1
outcome==1.3.0.post0
packaging==23.2
protobuf==4.24.4
psycopg2==2.9.9
pycparser==2.21
PyJWT==2.8.0
PySocks==1.7.1
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2023.4
requests==2.31.0
requests-oauthlib==1.3.1
selenium==4.14.0
sniffio==1.3.0
social-auth-app-django==5.4.0
social-auth-core==4.5.2
sortedcontainers==2.4.0
sqlparse==0.4.4
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.11.1
typing_extensions==4.9.0
tzdata==2023.3
urllib3==2.0.7
wsproto==1.2.0
yarl==1.9.4
goto-label==0.1.2
asteval==0.9.31
google-ai-generativelanguage==0.4.0
google-api-core==2.16.2
google-auth==2.27.0
google-generativeai==0.3.2
googleapis-common-protos==1.62.0
pillow==10.2.0
numpy==1.26.3
matplotlib==3.8.2